+++ sdk/core/src/dto/customMessageTemplates.dto.ts
@@ -1,0 +1,9 @@
+export class CreateMessageTemplateInput {
+  username: string
+  template: string
+  subject?: string
+  htmlTemplate?: string
+}
+export class DeleteMessageTemplateInput {
+  username: string
+}