+++ sdk/react-native/CHANGELOG.md
@@ -1,7 +1,12 @@
+# release 4.2.4 (2021-05-04)
+* custom messages support for passwordless login
+* vc-data winner schema
+
 # release 4.2.3 (2020-04-01)
  * add optional pagination to `AffinityWallet.getCredentials`
  * add new method `AffinityWallet.getCredentialByIndex` that returns credentials given at the index
  * fix `AffinityWallet.signUp` returning CommonNetworkMember instead of React Native AffinityWallet
+
 # release 4.2.2 
  * use new `vc-data`
 # release 4.2.1 (2020-02-22)